Suite aux résultats de test.js: - meilleurs résultats avec GrilleMatrice que GrilleListe - changements effectués: Grille.js remplacé par GrilleMatrice.js - suppression de coords dans Pixel.js pour cause de redondance.
