+++ src/Pixel.js
@@ -1,4 +1,3 @@
-import { Coordonnees } from "./Coordonnees.js";
 import { Couleur } from "./Couleur.js";
 
 /**
@@ -6,10 +5,6 @@ import { Couleur } from "./Couleur.js";
  */
 export class Pixel {
     // ATTRIBUTS
-    /**
-     * Coordonnées (x, y) du Pixel
-     */
-    #_coords;
     /**
      * Couleur du Pixel.
      */
@@ -25,16 +20,10 @@ export class Pixel {
      *  une couleur et est déselectionné.
      */
     constructor() {
-        this.#_coords = new Coordonnees();
         this.#_estSelectionne = false;
     }
 
     // GETTERS & SETTERS
-    /**
-     * Obtient les coordonnées du Pixel.
-     * @returns {Coordonnees} Les coordonnées du Pixel.
-     */
-    getCoord() { return this.#_coords; }
     /**
      * Obtient la couleur du Pixel.
      * @returns {Couleur} La couleur du Pixel.
@@ -46,11 +35,6 @@ export class Pixel {
      */
     isSelected() { return this.#_estSelectionne; }
 
-    /**
-     * Définit les coordonnées du Pixel par val.
-     * @param {Coordonnees} val Les nouvelles coordonnées du Pixel.
-     */
-    setCoord(val) { this.#_coords = val; }
     /**
      * Définit la couleur du Pixel.
      * @param {Couleur} c La nouvelle couleur du Pixel.
